Amplitudeinstall

+++ db/ODS/views/vw_dim_user.sql
@@ -167,7 +167,7 @@ inner join
 on user_dates.id = u.id
 left join 
   app_network an
-on u.id = an.user_id
+on u.id = an.user_id::integer
 ;
 
 -- select data_nascimento from vw_dim_user
